moved constants
