+++ examples/curated-paranet-demo.js
@@ -5,7 +5,7 @@ import {
     PARANET_NODES_ACCESS_POLICY,
     PARANET_MINERS_ACCESS_POLICY,
     PARANET_KC_SUBMISSION_POLICY,
-} from '../constants.js';
+} from '../constants/constants.js';
 
 const ENVIRONMENT = 'development';
 const OT_NODE_HOSTNAME = 'http://localhost';
